requirements.txt
asgiref==3.3.0
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3
cryptography==3.2.1
defusedxml==0.7.0rc1
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.1
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.4
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
rest-social-auth==4.2.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
social-auth-app-django==3.4.0
social-auth-core==3.3.3
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
